fix: rename test class for use in build process
